checkpoint: dbmw 冲突截断后现状 + 全服务端既有翻译成果
- source/dbmw/ManagerTypes.cpp 被并行 agent 的宏合并脚本截断（7714 行，
  缺框架/处理器实现），先提交当前状态作为恢复点
- 其余服务端（auction/monitor/manager/community 等）既有翻译成果
- 保留 dnf_installer gitlink 不变

+++ dnf_decompile/source/shared/common/include/DBTransactionDesign.h
@@ -158,7 +158,8 @@ struct ROI_AverageKey
     ROI_Category option_category;              // @4
     union
     {
-        __int64 option_index_key;              // @16
+        // ORIG DWARF：option_index_key 是 signed long long（非 __int64 typedef）
+        long long option_index_key;            // @16
         struct
         {
             short option_index_value[3];
@@ -175,11 +176,12 @@ struct ROI_AverageKey
     {
         if (baseItem_index == _rhp.baseItem_index)
         {
-            if ((unsigned long long)option_index_key == (unsigned long long)_rhp.option_index_key)
+            // ORIG：64 位比较为有符号（高字 jl/jg + 低字 jb），不要无符号强转
+            if (option_index_key == _rhp.option_index_key)
             {
                 return option_category < _rhp.option_category;
             }
-            return (unsigned long long)option_index_key < (unsigned long long)_rhp.option_index_key;
+            return option_index_key < _rhp.option_index_key;
         }
         return baseItem_index < _rhp.baseItem_index;
     }
